moved table body results into a new component
