+++ client/src/components/Search.jsx
@@ -80,3 +80,20 @@ export class Search extends Component {
         );
     }
 }
+
+/*render (){
+import ListItem from ''...
+
+    const styleObj = {...}
+    return (
+        <table>
+            {list.map(item => {
+                <ListItem 
+                data={item}
+                key={item}
+                index={i}
+            })}
+        </table>
+    )
+
+}*/
